fix: homepage layout
