+++ client/src/components/Card/Card.module.scss
@@ -1,8 +1,8 @@
 @import '../../mixins.scss';
 
 .card {
-  width: 90px;
-  height: 90px;
+  width: 110px;
+  height: 110px;
   border-radius: 10px;
   border: 1px solid #e8c4ba;
   background: #f0cec5;
