+++ client/src/components/GameInfo/GameInfo.module.scss
@@ -1,15 +1,22 @@
 @import '../../mixins.scss';
 
 .info {
+  width: 100%;
   .panel {
     display: flex;
     align-items: center;
     justify-content: space-between;
+    column-gap: 15px;
     padding: 15px;
     border-radius: 15px;
-    background: #baa6a3;
+    background: $dark;
     box-shadow: 0px 0px 20px -5px rgba(0, 0, 0, 0.25) inset;
     .meta {
+      color: $white;
+      flex: 1;
+      .time,
+      .movesCount {
+      }
       span {
         font-weight: 700;
       }
