+++ client/src/index.scss
@@ -28,7 +28,7 @@ body {
     min-height: 100vh;
   }
   main {
-    max-width: 435px;
+    width: 1200px;
     margin: 0 auto;
     height: 100%;
   }
