+++ client/src/pages/Homepage/Homepage.jsx
@@ -3,13 +3,17 @@ import s from './Homepage.module.scss';
 import Head from '../../components/Head/Head';
 import Field from '../../components/Field/Field';
 import GameInfo from '../../components/GameInfo/GameInfo';
+import Leaderboard from '../../components/Leaderboard/Leaderboard';
 
 const Homepage = () => {
   return (
     <div className={s.homepage}>
-      <Head />
-      <Field />
-      <GameInfo />
+      {/* <Head /> */}
+      <div className={s.game}>
+        <GameInfo />
+        <Field />
+      </div>
+      <Leaderboard />
     </div>
   );
 };
